refactor: use latest value ref hook
