+++ packages/curve-ui-kit/src/features/candle-chart/CandleChart.tsx
@@ -5,6 +5,7 @@ import { useEffect, useRef, useCallback, useMemo, type RefObject } from 'react'
 import { Box } from '@mui/material'
 import { maybes } from '@primitives/objects.utils'
 import { CHART_LINE_WIDTHS } from '@ui-kit/shared/ui/Chart/chart.utils'
+import { useLatestValueRef } from '../../hooks/useLatestValueRef'
 import { PRICE_SCALE_MARGINS } from './constants'
 import { createLiquidationRangeSeries } from './custom-series/liquidationRangeSeries'
 import type { LiquidationRangePoint, LiquidationRangeSeriesOptions } from './custom-series/liquidationRangeSeries'
@@ -12,7 +13,6 @@ import { useCandleTimeScaleSubscriptions } from './hooks/useCandleTimeScaleSubsc
 import type { ChartColors } from './hooks/useChartPalette'
 import { useHistoricalChartPagination } from './hooks/useHistoricalChartPagination'
 import { useInitialChartRightOffset } from './hooks/useInitialChartRightOffset'
-import { useLatestValueRef } from './hooks/useLatestValueRef'
 import { useVisiblePriceRangeSync } from './hooks/useVisiblePriceRangeSync'
 import type { LpPriceOhlcDataFormatted, OraclePriceData, LiquidationRanges, LlammaLiquididationRange } from './types'
 import { calculateRobustPriceRange, priceFormatter } from './utils'
